Added polylabel article, updated README
Article copied from the original post on the Mapbox blog:
https://blog.mapbox.com/a-new-algorithm-for-finding-a-visual-center-of-a-polygon-7c77e6492fbc
This includes a new directory, images/, to hold the article figures.
Some very light editing has also been done.

Also updated the project README to reference the article. Cleaned up the
formatting a bit to be useable in plaintext form.